Move function syntax related exceptions to package Exceptions.FunctionSyntax
