+++ ConfigCommands-core/src/main/java/me/willkroboth/ConfigCommands/Exceptions/FunctionSyntax/InvalidDoCommand.java
@@ -1,4 +1,4 @@
-package me.willkroboth.ConfigCommands.Exceptions;
+package me.willkroboth.ConfigCommands.Exceptions.FunctionSyntax;
 
 public class InvalidDoCommand extends InvalidExpressionCommand {
     public InvalidDoCommand(String arg, String reason) {
